chore: add TaskRow component & add controlled states updates in TaskContext
