+++ frontend/src/components/TaskToolbar.jsx
@@ -1,15 +1,12 @@
-import { useTaskContext } from '../context/TaskContext';
-import TaskForm from "./TaskForm"
+import TaskForm from './TaskForm'
 
 
 export default function TaskToolbar() {
 
-  const { fetchTasks } = useTaskContext();
-
   return(
     <div className="flex justify-between items-center mb-4">
 
-      <TaskForm onTaskAdded={fetchTasks} />
+      <TaskForm />
 
     </div>
   );
